fixed integration submit test

+++ ol/tower/tests/integration-submit-tx.rs
@@ -214,7 +214,7 @@ fn check_node_sync(tx_params: &TxParams, config: &AppCfg) -> Result<(), Error> {
     println!("Local tower height: {}", current_block_number);
 
     // The client can be in sync with local or -1 wrt local. 
-    if (current_block_number != remote_height) && (current_block_number - 1)!= remote_height {
+    if (i128::from(current_block_number) != remote_height) && i128::from(current_block_number - 1)!= remote_height {
         bail!("Block heights don't match: Miner: {}, Remote: {}", current_block_number, remote_height)
     }
     Ok(())
